Create app.js
Throug h traial and error, I wanted to achieve the typing effect, commented out the function for the timer code, spent a lot of time trying to create a p element on the fly while being stored letter by letter in the p tag Even now it remains as a text node and not inside the element node. Will ry something different to get the desired outcome. The code is stable for now.