Read description: setTimout and setInterval limitations and a bit about the event queue, execution stack
The biggest issue I was trying to solve with my limited knowlwdge was for a setTimout function to execute and wait before another setTimout function has finished but it wouldn't work. I have to  learn promises apparently do that.

The goal was to use opacity 0 to 1 with a setTimeout function, but since opacity still allocates space, i needed the divs to be deleted, but if I used setInterval/Timeout, the execution stack would not wait ironically; user could keep tapping the button and some divs would be deleted instantly since it was using the previous created divs delete timer and I was using forEach on the node list so everything would get deleted. I'll probably disable  the button intead if an error occurs for 2 seconds, that way user cannot spam the button when no inputs have been inserted.

tldr; execution stack has a higher priority than the events. that's why when i called the error div function to create a div, it would get deleted instantly depite setting 10000 millesecconds if spammed.